moved to ci and imagetagging
