+++ pkg/pipeline/CiHandler.go
@@ -29,7 +29,6 @@ import (
 	bean2 "github.com/devtron-labs/devtron/api/bean"
 	"github.com/devtron-labs/devtron/client/gitSensor"
 	"github.com/devtron-labs/devtron/internal/sql/repository/appWorkflow"
-	repository2 "github.com/devtron-labs/devtron/internal/sql/repository/imageTagging"
 	"github.com/devtron-labs/devtron/pkg/cluster"
 	repository3 "github.com/devtron-labs/devtron/pkg/cluster/repository"
 	k8s2 "github.com/devtron-labs/devtron/pkg/k8s"
@@ -65,8 +64,8 @@ type CiHandler interface {
 	CheckAndReTriggerCI(workflowStatus v1alpha1.WorkflowStatus) error
 	FetchMaterialsByPipelineId(pipelineId int, showAll bool) ([]pipelineConfig.CiPipelineMaterialResponse, error)
 	FetchMaterialsByPipelineIdAndGitMaterialId(pipelineId int, gitMaterialId int, showAll bool) ([]pipelineConfig.CiPipelineMaterialResponse, error)
-	FetchWorkflowDetails(appId int, pipelineId int, buildId int) (WorkflowResponse, error)
-	FetchArtifactsForCiJob(buildId int) (*ArtifactsForCiJob, error)
+	FetchWorkflowDetails(appId int, pipelineId int, buildId int) (types.WorkflowResponse, error)
+	FetchArtifactsForCiJob(buildId int) (*types.ArtifactsForCiJob, error)
 	//FetchBuildById(appId int, pipelineId int) (WorkflowResponse, error)
 	CancelBuild(workflowId int) (int, error)
 
@@ -74,14 +73,14 @@ type CiHandler interface {
 	GetHistoricBuildLogs(pipelineId int, workflowId int, ciWorkflow *pipelineConfig.CiWorkflow) (map[string]string, error)
 	//SyncWorkflows() error
 
-	GetBuildHistory(pipelineId int, appId int, offset int, size int) ([]WorkflowResponse, error)
+	GetBuildHistory(pipelineId int, appId int, offset int, size int) ([]types.WorkflowResponse, error)
 	DownloadCiWorkflowArtifacts(pipelineId int, buildId int) (*os.File, error)
 	UpdateWorkflow(workflowStatus v1alpha1.WorkflowStatus) (int, error)
 
 	FetchCiStatusForTriggerView(appId int) ([]*pipelineConfig.CiWorkflowStatus, error)
 	FetchCiStatusForTriggerViewV1(appId int) ([]*pipelineConfig.CiWorkflowStatus, error)
 	RefreshMaterialByCiPipelineMaterialId(gitMaterialId int) (refreshRes *gitSensor.RefreshGitMaterialResponse, err error)
-	FetchMaterialInfoByArtifactId(ciArtifactId int, envId int) (*GitTriggerInfoResponse, error)
+	FetchMaterialInfoByArtifactId(ciArtifactId int, envId int) (*types.GitTriggerInfoResponse, error)
 	WriteToCreateTestSuites(pipelineId int, buildId int, triggeredBy int)
 	UpdateCiWorkflowStatusFailure(timeoutForFailureCiBuild int) error
 	FetchCiStatusForTriggerViewForEnvironment(request resourceGroup.ResourceGroupingRequest) ([]*pipelineConfig.CiWorkflowStatus, error)
@@ -147,86 +146,6 @@ func NewCiHandlerImpl(Logger *zap.SugaredLogger, ciService CiService, ciPipeline
 	return cih
 }
 
-type WorkflowResponse struct {
-	Id                   int                                         `json:"id"`
-	Name                 string                                      `json:"name"`
-	Status               string                                      `json:"status"`
-	PodStatus            string                                      `json:"podStatus"`
-	Message              string                                      `json:"message"`
-	StartedOn            time.Time                                   `json:"startedOn"`
-	FinishedOn           time.Time                                   `json:"finishedOn"`
-	CiPipelineId         int                                         `json:"ciPipelineId"`
-	Namespace            string                                      `json:"namespace"`
-	LogLocation          string                                      `json:"logLocation"`
-	BlobStorageEnabled   bool                                        `json:"blobStorageEnabled"`
-	GitTriggers          map[int]pipelineConfig.GitCommit            `json:"gitTriggers"`
-	CiMaterials          []pipelineConfig.CiPipelineMaterialResponse `json:"ciMaterials"`
-	TriggeredBy          int32                                       `json:"triggeredBy"`
-	Artifact             string                                      `json:"artifact"`
-	TriggeredByEmail     string                                      `json:"triggeredByEmail"`
-	Stage                string                                      `json:"stage"`
-	ArtifactId           int                                         `json:"artifactId"`
-	IsArtifactUploaded   bool                                        `json:"isArtifactUploaded"`
-	IsVirtualEnvironment bool                                        `json:"isVirtualEnvironment"`
-	PodName              string                                      `json:"podName"`
-	EnvironmentId        int                                         `json:"environmentId"`
-	EnvironmentName      string                                      `json:"environmentName"`
-	ImageReleaseTags     []*repository2.ImageTag                     `json:"imageReleaseTags"`
-	ImageComment         *repository2.ImageComment                   `json:"imageComment"`
-	AppWorkflowId        int                                         `json:"appWorkflowId"`
-	CustomTag            *bean2.CustomTagErrorResponse               `json:"customTag,omitempty"`
-	PipelineType         string                                      `json:"pipelineType"`
-	ReferenceWorkflowId  int                                         `json:"referenceWorkflowId"`
-}
-
-type ArtifactsForCiJob struct {
-	Artifacts []string `json:"artifacts"`
-}
-
-type GitTriggerInfoResponse struct {
-	CiMaterials      []pipelineConfig.CiPipelineMaterialResponse `json:"ciMaterials"`
-	TriggeredByEmail string                                      `json:"triggeredByEmail"`
-	LastDeployedTime string                                      `json:"lastDeployedTime,omitempty"`
-	AppId            int                                         `json:"appId"`
-	AppName          string                                      `json:"appName"`
-	EnvironmentId    int                                         `json:"environmentId"`
-	EnvironmentName  string                                      `json:"environmentName"`
-	Default          bool                                        `json:"default,omitempty"`
-	ImageTaggingData ImageTaggingResponseDTO                     `json:"imageTaggingData"`
-	Image            string                                      `json:"image"`
-}
-
-type Trigger struct {
-	PipelineId                int
-	CommitHashes              map[int]pipelineConfig.GitCommit
-	CiMaterials               []*pipelineConfig.CiPipelineMaterial
-	TriggeredBy               int32
-	InvalidateCache           bool
-	ExtraEnvironmentVariables map[string]string // extra env variables which will be used for CI
-	EnvironmentId             int
-	PipelineType              string
-	CiArtifactLastFetch       time.Time
-	ReferenceCiWorkflowId     int
-}
-
-func (obj Trigger) BuildTriggerObject(refCiWorkflow *pipelineConfig.CiWorkflow,
-	ciMaterials []*pipelineConfig.CiPipelineMaterial, triggeredBy int32,
-	invalidateCache bool, extraEnvironmentVariables map[string]string,
-	pipelineType string) {
-
-	obj.PipelineId = refCiWorkflow.CiPipelineId
-	obj.CommitHashes = refCiWorkflow.GitTriggers
-	obj.CiMaterials = ciMaterials
-	obj.TriggeredBy = triggeredBy
-	obj.InvalidateCache = invalidateCache
-	obj.EnvironmentId = refCiWorkflow.EnvironmentId
-	obj.ReferenceCiWorkflowId = refCiWorkflow.Id
-	obj.InvalidateCache = invalidateCache
-	obj.ExtraEnvironmentVariables = extraEnvironmentVariables
-	obj.PipelineType = pipelineType
-
-}
-
 const WorkflowCancel = "CANCELLED"
 const DefaultCiWorkflowNamespace = "devtron-ci"
 const Running = "Running"
@@ -281,7 +200,7 @@ func (impl *CiHandlerImpl) reTriggerCi(retryCount int, refCiWorkflow *pipelineCo
 		return err
 	}
 
-	trigger := Trigger{}
+	trigger := types.Trigger{}
 	trigger.BuildTriggerObject(refCiWorkflow, ciMaterials, 1, true, nil, "")
 	_, err = impl.ciService.TriggerCiPipeline(trigger)
 
@@ -310,7 +229,7 @@ func (impl *CiHandlerImpl) HandleCIManual(ciTriggerRequest bean.CiTriggerRequest
 		createdOn = ciArtifact.CreatedOn
 	}
 
-	trigger := Trigger{
+	trigger := types.Trigger{
 		PipelineId:                ciTriggerRequest.PipelineId,
 		CommitHashes:              commitHashes,
 		CiMaterials:               nil,
@@ -356,7 +275,7 @@ func (impl *CiHandlerImpl) HandleCIWebhook(gitCiTriggerRequest bean.GitCiTrigger
 		return 0, err
 	}
 
-	trigger := Trigger{
+	trigger := types.Trigger{
 		PipelineId:                ciPipeline.Id,
 		CommitHashes:              commitHashes,
 		CiMaterials:               ciMaterials,
@@ -556,7 +475,7 @@ func (impl *CiHandlerImpl) FetchMaterialsByPipelineId(pipelineId int, showAll bo
 	return ciPipelineMaterialResponses, nil
 }
 
-func (impl *CiHandlerImpl) GetBuildHistory(pipelineId int, appId int, offset int, size int) ([]WorkflowResponse, error) {
+func (impl *CiHandlerImpl) GetBuildHistory(pipelineId int, appId int, offset int, size int) ([]types.WorkflowResponse, error) {
 	ciMaterials, err := impl.ciPipelineMaterialRepository.GetByPipelineIdForRegexAndFixed(pipelineId)
 	if err != nil {
 		impl.Logger.Errorw("ciMaterials fetch failed", "err", err)
@@ -596,9 +515,9 @@ func (impl *CiHandlerImpl) GetBuildHistory(pipelineId int, appId int, offset int
 		return nil, err
 	}
 
-	var ciWorkLowResponses []WorkflowResponse
+	var ciWorkLowResponses []types.WorkflowResponse
 	for _, w := range workFlows {
-		wfResponse := WorkflowResponse{
+		wfResponse := types.WorkflowResponse{
 			Id:                  w.Id,
 			Name:                w.Name,
 			Status:              w.Status,
@@ -698,33 +617,33 @@ func (impl *CiHandlerImpl) CancelBuild(workflowId int) (int, error) {
 	return workflow.Id, nil
 }
 
-func (impl *CiHandlerImpl) FetchWorkflowDetails(appId int, pipelineId int, buildId int) (WorkflowResponse, error) {
+func (impl *CiHandlerImpl) FetchWorkflowDetails(appId int, pipelineId int, buildId int) (types.WorkflowResponse, error) {
 	workflow, err := impl.ciWorkflowRepository.FindById(buildId)
 	if err != nil {
 		impl.Logger.Errorw("err", "err", err)
-		return WorkflowResponse{}, err
+		return types.WorkflowResponse{}, err
 	}
 	triggeredByUser, err := impl.userService.GetById(workflow.TriggeredBy)
 	if err != nil && !util.IsErrNoRows(err) {
 		impl.Logger.Errorw("err", "err", err)
-		return WorkflowResponse{}, err
+		return types.WorkflowResponse{}, err
 	}
 
 	if workflow.CiPipeline.AppId != appId {
 		impl.Logger.Error("pipeline does not exist for this app")
-		return WorkflowResponse{}, errors.New("invalid app and pipeline combination")
+		return types.WorkflowResponse{}, errors.New("invalid app and pipeline combination")
 	}
 
 	ciMaterials, err := impl.ciPipelineMaterialRepository.GetByPipelineId(pipelineId)
 	if err != nil {
 		impl.Logger.Errorw("err", "err", err)
-		return WorkflowResponse{}, err
+		return types.WorkflowResponse{}, err
 	}
 
 	ciArtifact, err := impl.ciArtifactRepository.GetByWfId(workflow.Id)
 	if err != nil && !util.IsErrNoRows(err) {
 		impl.Logger.Errorw("err", "err", err)
-		return WorkflowResponse{}, err
+		return types.WorkflowResponse{}, err
 	}
 
 	var ciMaterialsArr []pipelineConfig.CiPipelineMaterialResponse
@@ -745,11 +664,11 @@ func (impl *CiHandlerImpl) FetchWorkflowDetails(appId int, pipelineId int, build
 		env, err := impl.envRepository.FindById(workflow.EnvironmentId)
 		if err != nil && err != pg.ErrNoRows {
 			impl.Logger.Errorw("error in fetching environment details ", "err", err)
-			return WorkflowResponse{}, err
+			return types.WorkflowResponse{}, err
 		}
 		environmentName = env.Name
 	}
-	workflowResponse := WorkflowResponse{
+	workflowResponse := types.WorkflowResponse{
 		Id:                 workflow.Id,
 		Name:               workflow.Name,
 		Status:             workflow.Status,
@@ -775,13 +694,13 @@ func (impl *CiHandlerImpl) FetchWorkflowDetails(appId int, pipelineId int, build
 	return workflowResponse, nil
 }
 
-func (impl *CiHandlerImpl) FetchArtifactsForCiJob(buildId int) (*ArtifactsForCiJob, error) {
+func (impl *CiHandlerImpl) FetchArtifactsForCiJob(buildId int) (*types.ArtifactsForCiJob, error) {
 	artifacts, err := impl.ciArtifactRepository.GetArtifactsByParentCiWorkflowId(buildId)
 	if err != nil {
 		impl.Logger.Errorw("error in fetching artifacts by parent ci workflow id", "err", err, "buildId", buildId)
 		return nil, err
 	}
-	artifactsResponse := &ArtifactsForCiJob{
+	artifactsResponse := &types.ArtifactsForCiJob{
 		Artifacts: artifacts,
 	}
 	return artifactsResponse, nil
@@ -1426,30 +1345,30 @@ func (impl *CiHandlerImpl) FetchCiStatusForTriggerView(appId int) ([]*pipelineCo
 	return ciWorkflowStatuses, nil
 }
 
-func (impl *CiHandlerImpl) FetchMaterialInfoByArtifactId(ciArtifactId int, envId int) (*GitTriggerInfoResponse, error) {
+func (impl *CiHandlerImpl) FetchMaterialInfoByArtifactId(ciArtifactId int, envId int) (*types.GitTriggerInfoResponse, error) {
 
 	ciArtifact, err := impl.ciArtifactRepository.Get(ciArtifactId)
 	if err != nil {
 		impl.Logger.Errorw("err", "ciArtifactId", ciArtifactId, "err", err)
-		return &GitTriggerInfoResponse{}, err
+		return &types.GitTriggerInfoResponse{}, err
 	}
 
 	ciPipeline, err := impl.ciPipelineRepository.FindById(ciArtifact.PipelineId)
 	if err != nil {
 		impl.Logger.Errorw("err", "ciArtifactId", ciArtifactId, "err", err)
-		return &GitTriggerInfoResponse{}, err
+		return &types.GitTriggerInfoResponse{}, err
 	}
 
 	ciMaterials, err := impl.ciPipelineMaterialRepository.GetByPipelineId(ciPipeline.Id)
 	if err != nil {
 		impl.Logger.Errorw("err", "err", err)
-		return &GitTriggerInfoResponse{}, err
+		return &types.GitTriggerInfoResponse{}, err
 	}
 
 	deployDetail, err := impl.appListingRepository.DeploymentDetailByArtifactId(ciArtifactId, envId)
 	if err != nil {
 		impl.Logger.Errorw("err", "err", err)
-		return &GitTriggerInfoResponse{}, err
+		return &types.GitTriggerInfoResponse{}, err
 	}
 
 	ciMaterialsArr := make([]pipelineConfig.CiPipelineMaterialResponse, 0)
@@ -1461,20 +1380,20 @@ func (impl *CiHandlerImpl) FetchMaterialInfoByArtifactId(ciArtifactId int, envId
 			workflow, err = impl.ciWorkflowRepository.FindLastTriggeredWorkflowByArtifactId(ciArtifact.ParentCiArtifact)
 			if err != nil {
 				impl.Logger.Errorw("err", "ciArtifactId", ciArtifact.ParentCiArtifact, "err", err)
-				return &GitTriggerInfoResponse{}, err
+				return &types.GitTriggerInfoResponse{}, err
 			}
 		} else {
 			workflow, err = impl.ciWorkflowRepository.FindLastTriggeredWorkflowByArtifactId(ciArtifactId)
 			if err != nil {
 				impl.Logger.Errorw("err", "ciArtifactId", ciArtifactId, "err", err)
-				return &GitTriggerInfoResponse{}, err
+				return &types.GitTriggerInfoResponse{}, err
 			}
 		}
 
 		triggeredByUser, err = impl.userService.GetById(workflow.TriggeredBy)
 		if err != nil && !util.IsErrNoRows(err) {
 			impl.Logger.Errorw("err", "err", err)
-			return &GitTriggerInfoResponse{}, err
+			return &types.GitTriggerInfoResponse{}, err
 		}
 
 		for _, m := range ciMaterials {
@@ -1522,9 +1441,9 @@ func (impl *CiHandlerImpl) FetchMaterialInfoByArtifactId(ciArtifactId int, envId
 	imageTaggingData, err := impl.imageTaggingService.GetTagsData(ciPipeline.Id, ciPipeline.AppId, ciArtifactId, false)
 	if err != nil {
 		impl.Logger.Errorw("error in fetching imageTaggingData", "err", err, "ciPipelineId", ciPipeline.Id, "appId", ciPipeline.AppId, "ciArtifactId", ciArtifactId)
-		return &GitTriggerInfoResponse{}, err
+		return &types.GitTriggerInfoResponse{}, err
 	}
-	gitTriggerInfoResponse := &GitTriggerInfoResponse{
+	gitTriggerInfoResponse := &types.GitTriggerInfoResponse{
 		//GitTriggers:      workflow.GitTriggers,
 		CiMaterials:      ciMaterialsArr,
 		TriggeredByEmail: triggeredByUser.EmailId,
