+++ pkg/pipeline/types/CiCdConfig.go
@@ -14,6 +14,7 @@ import (
 	"os/user"
 	"path/filepath"
 	"strings"
+	"time"
 )
 
 type CiCdConfig struct {
@@ -472,3 +473,51 @@ const BLOB_STORAGE_S3 = "S3"
 const BLOB_STORAGE_GCP = "GCP"
 
 const BLOB_STORAGE_MINIO = "MINIO"
+
+type ArtifactsForCiJob struct {
+	Artifacts []string `json:"artifacts"`
+}
+
+type GitTriggerInfoResponse struct {
+	CiMaterials      []pipelineConfig.CiPipelineMaterialResponse `json:"ciMaterials"`
+	TriggeredByEmail string                                      `json:"triggeredByEmail"`
+	LastDeployedTime string                                      `json:"lastDeployedTime,omitempty"`
+	AppId            int                                         `json:"appId"`
+	AppName          string                                      `json:"appName"`
+	EnvironmentId    int                                         `json:"environmentId"`
+	EnvironmentName  string                                      `json:"environmentName"`
+	Default          bool                                        `json:"default,omitempty"`
+	ImageTaggingData ImageTaggingResponseDTO                     `json:"imageTaggingData"`
+	Image            string                                      `json:"image"`
+}
+
+type Trigger struct {
+	PipelineId                int
+	CommitHashes              map[int]pipelineConfig.GitCommit
+	CiMaterials               []*pipelineConfig.CiPipelineMaterial
+	TriggeredBy               int32
+	InvalidateCache           bool
+	ExtraEnvironmentVariables map[string]string // extra env variables which will be used for CI
+	EnvironmentId             int
+	PipelineType              string
+	CiArtifactLastFetch       time.Time
+	ReferenceCiWorkflowId     int
+}
+
+func (obj Trigger) BuildTriggerObject(refCiWorkflow *pipelineConfig.CiWorkflow,
+	ciMaterials []*pipelineConfig.CiPipelineMaterial, triggeredBy int32,
+	invalidateCache bool, extraEnvironmentVariables map[string]string,
+	pipelineType string) {
+
+	obj.PipelineId = refCiWorkflow.CiPipelineId
+	obj.CommitHashes = refCiWorkflow.GitTriggers
+	obj.CiMaterials = ciMaterials
+	obj.TriggeredBy = triggeredBy
+	obj.InvalidateCache = invalidateCache
+	obj.EnvironmentId = refCiWorkflow.EnvironmentId
+	obj.ReferenceCiWorkflowId = refCiWorkflow.Id
+	obj.InvalidateCache = invalidateCache
+	obj.ExtraEnvironmentVariables = extraEnvironmentVariables
+	obj.PipelineType = pipelineType
+
+}
