+++ pkg/pipeline/types/Workflow.go
@@ -24,6 +24,7 @@ import (
 	"github.com/devtron-labs/common-lib/blob-storage"
 	bean3 "github.com/devtron-labs/devtron/api/bean"
 	repository2 "github.com/devtron-labs/devtron/internal/sql/repository"
+	repository3 "github.com/devtron-labs/devtron/internal/sql/repository/imageTagging"
 	"github.com/devtron-labs/devtron/internal/sql/repository/pipelineConfig"
 	bean2 "github.com/devtron-labs/devtron/pkg/bean"
 	"github.com/devtron-labs/devtron/pkg/cluster/repository"
@@ -581,4 +582,36 @@ type AppLabelMetadata struct {
 type ChildCdMetadata struct {
 	ChildCdEnvName     string `json:"CHILD_CD_ENV_NAME"`
 	ChildCdClusterName string `json:"CHILD_CD_CLUSTER_NAME"`
+}
+
+type WorkflowResponse struct {
+	Id                   int                                         `json:"id"`
+	Name                 string                                      `json:"name"`
+	Status               string                                      `json:"status"`
+	PodStatus            string                                      `json:"podStatus"`
+	Message              string                                      `json:"message"`
+	StartedOn            time.Time                                   `json:"startedOn"`
+	FinishedOn           time.Time                                   `json:"finishedOn"`
+	CiPipelineId         int                                         `json:"ciPipelineId"`
+	Namespace            string                                      `json:"namespace"`
+	LogLocation          string                                      `json:"logLocation"`
+	BlobStorageEnabled   bool                                        `json:"blobStorageEnabled"`
+	GitTriggers          map[int]pipelineConfig.GitCommit            `json:"gitTriggers"`
+	CiMaterials          []pipelineConfig.CiPipelineMaterialResponse `json:"ciMaterials"`
+	TriggeredBy          int32                                       `json:"triggeredBy"`
+	Artifact             string                                      `json:"artifact"`
+	TriggeredByEmail     string                                      `json:"triggeredByEmail"`
+	Stage                string                                      `json:"stage"`
+	ArtifactId           int                                         `json:"artifactId"`
+	IsArtifactUploaded   bool                                        `json:"isArtifactUploaded"`
+	IsVirtualEnvironment bool                                        `json:"isVirtualEnvironment"`
+	PodName              string                                      `json:"podName"`
+	EnvironmentId        int                                         `json:"environmentId"`
+	EnvironmentName      string                                      `json:"environmentName"`
+	ImageReleaseTags     []*repository3.ImageTag                     `json:"imageReleaseTags"`
+	ImageComment         *repository3.ImageComment                   `json:"imageComment"`
+	AppWorkflowId        int                                         `json:"appWorkflowId"`
+	CustomTag            *bean3.CustomTagErrorResponse               `json:"customTag,omitempty"`
+	PipelineType         string                                      `json:"pipelineType"`
+	ReferenceWorkflowId  int                                         `json:"referenceWorkflowId"`
 }
